+++ packages/node/src/indexer/worker/worker-fetch.module.ts
@@ -21,6 +21,7 @@ import { ProjectService } from '../project.service';
 import { WorkerRuntimeService } from '../runtime/workerRuntimeService';
 import { SandboxService } from '../sandbox.service';
 import { UnfinalizedBlocksService } from '../unfinalizedBlocks.service';
+import { WorkerInMemoryCacheService } from './worker.inMemoryCache.service';
 import { WorkerService } from './worker.service';
 import { WorkerUnfinalizedBlocksService } from './worker.unfinalizedBlocks.service';
 
@@ -78,7 +79,10 @@ import { WorkerUnfinalizedBlocksService } from './worker.unfinalizedBlocks.servi
     },
     WorkerService,
     WorkerRuntimeService,
-    InMemoryCacheService,
+    {
+      provide: InMemoryCacheService,
+      useFactory: () => new WorkerInMemoryCacheService((global as any).host),
+    },
   ],
   exports: [],
 })
