+++ packages/types-core/src/interfaces.ts
@@ -4,6 +4,6 @@
 export type DynamicDatasourceCreator = (name: string, args: Record<string, unknown>) => Promise<void>;
 
 export interface Cache {
-  set<D = string>(key: string, value: D): Promise<void>;
-  get<D = string>(key: string): Promise<D | undefined>;
+  set<D>(key: string, value: D): Promise<void>;
+  get<D>(key: string): Promise<D | undefined>;
 }
